move func update
